fix: change note on indic2&3 and add redirect between 2, 3, 2&3
